move all components to src, including provider and theme + rewrite button story
