+++ apps/storybook/stories/components/Button/Button.mdx
@@ -1,0 +1,24 @@
+import { Title, Canvas, Meta, Story, Controls } from "@storybook/blocks";
+import * as ButtonStories from "./Button.stories";
+
+<Meta of={ButtonStories} />
+
+# Button
+
+Her er en oversikt over utvalget av Knapper i KVIB. Disse er blant annet bygget opp ved bruk av desing tokens fra KVIB, eksempelvis farger og typografi. I dokumentasjonen følger beskrivelse av komponenten og bruksområde, sammen med kodeeksempler.
+
+Vi har to sett med farger. Vi har et primærsett (grønn) og et sekundærsett (blått). Vår palett er inspirert av jordtoner. Vi har valgt sekundærsettet som er blått på karttjenester og forvaltningstjenester for å ha kontrast mot kartet.
+
+Løsninger som skal se ut som en del av [kartverket.no](http://kartverket.no) skal bruke grønn.
+
+<Canvas of={ButtonStories.Button} />
+
+## Ta i bruk
+
+```jsx
+import { Button } from "@kvib/react";
+```
+
+## Props
+
+<Controls of={ButtonStories.Button} />
