+++ packages/react/src/provider/index.tsx
@@ -1,0 +1,1 @@
+export * from "./KvibProvider";
